improve extensibility of mass redistribution method in axial expansion
